requirements.txt
Django==4.0.5
django-durationwidget==1.0.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
Pillow==9.2.0

django-environ==0.10.0
psycopg2==2.9.5

celery==5.2.7
django-celery-results==2.5.0
django-celery-beat==2.5.0

django-filter==23.1
django-extensions==3.2.1
django-storages==1.13.2

drf-spectacular==0.26.1

django-redis==5.2.0
